Allowing global and metric-specific dimensions to coexist
DimensionMapper copied all global dimensions into each metric-specific
dimension list. Subsequent calls to getDimensions() would loop through
the global dimensions and then the metric-specific dimensions (which now
contain the global dimensions), adding each to an ArrayList, which
caused each global dimension to be duplicated in the list.

CloudWatch throws an exception if a single dimension is defined more
than once on a metric, which leads to the CloudWatchMetricRecorder
shutting down without this change.