Explicitly aggregate counts within a context in MetricRecorder
While the javadocs previously indicated that Counts were to be explicitly aggregated values (where only the total at the end of a context had any meaning), it was left up to MetricRecorder implementations to get this right.  This change enforces the contract from the javadoc by handling the aggregation of counts in the abstract MetricRecorder class.