Upgrade to AWS SDK 2, JUnit 5
Move from AWS SDK 1 to AWS SDK 2 and swap to JUnit 5.
AWS SDK 2 moves away from using the bundle of AWS SDKs for all AWS services.

There will be a breaking change for consumers of Swage Metrics Core, since AWS SDK 2 brings in different namespacing and a new CloudWatch client.  We're currently only doing a snapshot version so we didn't bump it to 1.0 or 2.0-SNAPSHOT.